New version
- Update package dependencies and lock file version.
- Skip closing if the connection is not established.
- Deprecate outdated node version.

+++ src/services/engine.js
@@ -149,7 +149,9 @@ module.exports = class EngineService extends EventEmitter {
   close() {
     return lock.unlock(this._getLockPath())
       .finally(() => {
-        this._process.kill();
+        if (this._process) {
+          this._process.kill();
+        }
       });
   }
 };
